Added file existence checker in PinataCRUD Operations v.3.0

+++ PinataCRUD/DeleteOperation.py
@@ -1,21 +1,21 @@
 import requests
 from PinataCredentials import credential
+import FileExistenceChecker
 
 # Function to delete file on IPFS via Pinata
 def delete_from_ipfs(ipfs_hash):
 
-    response = requests.delete(f"https://api.pinata.cloud/pinning/unpin/{ipfs_hash}", headers=credential)
-
-    if response.status_code == 200:
-        return True
+    if FileExistenceChecker.check_file_existence(ipfs_hash):
+        
+        response = requests.delete(f"https://api.pinata.cloud/pinning/unpin/{ipfs_hash}", headers=credential)
+        if response.status_code == 200:
+            print("File deleted from IPFS")
+            return True
     else:
-        print("Error:", response.text)
+        print("Error file can't be deleted from IPFS")
         return False
 
-
-
-##ipfs_hash = "QmYVAAfnqtEMqBqRggma24kSFY51CZM61JrzYPz9SJVL5m"
-##deleted = delete_from_ipfs(ipfs_hash)
+##deleted = delete_from_ipfs("QmRSDGXdZx7PZbLAWMT39xH9HWAnXHHMvHUhLC7VmMoA9h")
 ##if deleted:
 ##    print("File deleted from IPFS")
 ##else:
